requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-cloud-storage==2.13.0
google-cloud-aiplatform==1.38.1
python-dotenv==1.0.0
pydantic==2.5.2
